% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockRandomization.R
\name{blockRandomization}
\alias{blockRandomization}
\title{Block Randomization of Treatment Assignment for Experiments}
\usage{
blockRandomization(data, n = NULL, block = NULL, sort = TRUE)
}
\arguments{
\item{data}{A data frame for which the treatments are randomly assigned.}

\item{n}{Number of observations from each block to be randomized into the treatment group. 
Must be the same length as the number of unique values in \code{block}.}

\item{block}{A vector which denotes which block each observation is in. Observations will
be randomized within these blocks.}

\item{sort}{If \code{TRUE}, returns data in original order. If \code{FALSE}, returns data
grouped by the categories defined by block.}
}
\value{
A list of class \code{blockRandomize} which contains the following items:
\item{call}{ the matched call. }
\item{data}{The data frame that was used to conduct the randomization.}
\item{treatment}{ The vector of randomized treatments.}
\item{block}{The block vector.}
}
\description{
This function implements block randomization of treatment assignment for
randomized experiments. Complete randomization is conducted within blocks 
defined by a variable provided by the user.
}
\examples{
## create sample data
block <- rep(c("A", "B", "A", "C"), times = c(15,20,5,20))
data <- data.frame(x1 = rnorm(60), x2 = rnorm(60), block = block)

## assigns 10, 5, and 5 to treatment in groups A, B, C
blockRandomization(data, n = c(10,5,5), block = data$block)

}
\author{
Kosuke Imai, Professor of Government and Statistics, Harvard University
\email{imai@harvard.edu}

Tyler Simko, PhD Student, Harvard University
\email{tsimko@g.harvard.edu}
}
