% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{browserAddin}
\alias{browserAddin}
\title{RStudio AddIn function.}
\usage{
browserAddin(steps = fullhistory())
}
\arguments{
\item{steps}{A \code{steps} data structure, see \link{fullhistory} for an example.}
}
\description{
\code{browserAddin} runs the \code{htmlwidget} implemented in this
package as a RStudio AddIn (see \link[shiny:runGadget]{shiny::runGadget} for details).
}
