% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.r
\name{checkout_commit}
\alias{checkout_commit}
\title{Restore state of work from before into global environment.}
\usage{
checkout_commit(id, clear = TRUE)
}
\arguments{
\item{id}{Commit identifier.}

\item{clear}{Clear global environment before restoring the commit.}
}
\description{
Restore state of work from before into global environment.
}

