% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{cleanup_object}
\alias{cleanup_object}
\title{Removes references to environments.}
\usage{
cleanup_object(obj)
}
\arguments{
\item{obj}{Object to be processed.}
}
\value{
\code{obj} with environment references replaced by \code{emptyenv()}
}
\description{
Some objects (e.g. formula, lm) store references to environments
in which they were created. This function replaces each such reference
with a reference to \code{emptyenv()}.
}
