% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{combine_dots}
\alias{combine_dots}
\title{Combine ellipsis with optional list of dots.}
\usage{
combine_dots(dots, dots2)
}
\arguments{
\item{dots}{Dots created with \link[lazyeval]{\code{lazy_dots}}.}

\item{dots2}{Another (optional) \emph{lazy dots} or a plain list;
passed to \link[lazyeval]{\code{as.lazy_dots}}.}
}
\value{
A combined \code{list} of lazy objects.
}
\description{
Some functions allow for passing both ellipsis argument which they
convert to a \code{lazy_dots} object, and a \code{list} that needs
to be combined with ellipsis. This happens when \emph{lazy dots}
need to be passed between user-facing functions.
}

