% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{commit}
\alias{commit}
\alias{print.commit}
\title{Creates a new commit object.}
\usage{
commit(contents, expression, parent, id, object_ids)

\method{print}{commit}(x, simple = FALSE, ...)
}
\description{
Creates a new commit object.
}
