% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{commit_store}
\alias{commit_store}
\title{Write commit to an object store.}
\usage{
commit_store(commit, store)
}
\arguments{
\item{commit}{A [commit] object.}

\item{store}{An object store, e.g. [storage::filesystem].}
}
\description{
Write commit to an object store.
}
