% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_models}
\alias{compare_models}
\alias{compare_models.default}
\alias{compare_models.clist}
\alias{compare_models.list}
\title{Compare models.}
\usage{
compare_models(x, ...)

\method{compare_models}{default}(x, ...)

\method{compare_models}{clist}(x, ...)

\method{compare_models}{list}(x, ...)
}
\description{
Compare models.
}
