% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.r
\name{create_commit}
\alias{create_commit}
\title{Create a commit object.}
\usage{
create_commit(env, storage)
}
\arguments{
\item{env}{Environment to create this commit from.}

\item{storage}{Storage where newly created objects (if any) will be stored.}
}
\value{
A \code{commit} object.
}
\description{
Create a commit object.
}

