% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.r
\name{extract_assignment}
\alias{extract_assignment}
\title{Extract RHS of assignment to a variable.}
\usage{
extract_assignment(name, expression)
}
\arguments{
\item{name}{Variable name to search for.}

\item{expression}{Any expression where \code{name} is expected to be
assigned.}
}
\value{
RHS of the assignment expression.
}
\description{
Extract RHS of assignment to a variable.
}

