% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.r
\name{extract_parents}
\alias{extract_parents}
\title{Identify objects referenced in an expression.}
\usage{
extract_parents(expression, env)
}
\arguments{
\item{expression}{Expression to process.}

\item{env}{Environment to search for objects in.}
}
\value{

}
\description{
Finds all objects from \code{env} referenced in \code{expression}.
}

