% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_to_steps}
\alias{graph_to_steps}
\title{Transform a graph of commits into a graph of steps.}
\usage{
graph_to_steps(graph)
}
\arguments{
\item{graph}{Object returned by [graph()].}
}
\value{
Object of S3 class `steps`.
}
\description{
A step is an introduction of a new object or a new plot to the
session. Graph of steps is easier to read for humans than a graph of
commits becase only the relevant (new) information is shown in each
node of the graph. Thus, translating from commits to steps is the
first step to visualize the history stored in commits.
}
