% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_to_tree}
\alias{graph_to_tree}
\title{Turn a graph of \code{commit}s into a tree.}
\usage{
graph_to_tree(g)
}
\description{
Used mainly when preparing a printout of a sequence or actual tree of
\code{commit}s.
}
