% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{internal_object_store}
\alias{internal_object_store}
\alias{discover_object_store}
\alias{prepare_object_store}
\title{Manage object stores.}
\usage{
discover_object_store(path = getwd())

prepare_object_store(path, silent = !interactive())
}
\arguments{
\item{path}{Path to be examined.}
}
\value{
\code{discover_object_store} returns a \code{character} vector whose
elemnets are paths to existing object stores.

\code{prepare_object_store} returns an \link[storage:object_store]{storage::object_store} object.
}
\description{
Manage object stores.

\code{discover_object_store} starts with with given \code{path}
and searches for a configuration file (\emph{not implemented yet}) or an
existing object store that \code{experiment} can use for the current R
session. If a store can be found, an \link[storage:object_store]{storage::object_store} object
is returned; otherwise, a \code{NULL} value is returned.

\code{prepare_object_store} opens an existing object store
(via \code{discover_object_store}) or creates a new, temporary one.
}
