% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.r
\name{is_assignment}
\alias{is_assignment}
\title{Returns \code{TRUE} if an assignment expression.}
\usage{
is_assignment(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
\code{TRUE} is \code{x} is an assignment, \code{FALSE} otherwise.
}
\description{
Identifies three types of assignments: \code{<-}, \code{<<-} and
\code{=}. Returns \code{TRUE} is \code{x} is a \code{call} and
its first element is identical with either of the three.
}

