% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.r
\name{is_commit}
\alias{is_commit}
\title{Does an object inherit from \code{commit}.}
\usage{
is_commit(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\value{
\code{TRUE} if \code{x} is a commit.
}
\description{
Does an object inherit from \code{commit}.
}

