% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.r
\name{object_exists}
\alias{object_exists}
\title{Check whether an object exists in storage.}
\usage{
object_exists(st, id)
}
\arguments{
\item{st}{Storage to check.}

\item{id}{Object identifier.}
}
\value{
\code{TRUE} if object exists, \code{FALSE} otherwise.
}
\description{
Check whether an object exists in storage.
}

