% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{plot_as_svg}
\alias{plot_as_svg}
\title{Returns a base64-encoded, SVG plot.}
\usage{
plot_as_svg(pl)
}
\arguments{
\item{pl}{Plot recorded by \code{\link[=recordPlot]{recordPlot()}}.}
}
\value{
\code{character} string, base64-encoded SVG plot.
}
\description{
Returns a base64-encoded, SVG plot.
}
