% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_by}
\alias{query_by}
\title{Query the current history.}
\usage{
query_by(..., .related = "plots")
}
\arguments{
\item{...}{Search conditions.}

\item{.related}{Included related entities (objects or plots).}
}
\value{
A reduced history graph.
}
\description{
The following helper functions can be use in expressions to define
conditions:
* `is_named(...)` name matches any from the list
* `inherits` object inherits from any of the specified classes
}
\details{
The following variables can be used in expressions when defining
conditions:
* `name` object name
* `class` object class
* `id` object identifier`
}
\examples{
\dontrun{
# search for a specific class
query_by(inherits("lm", "data.frame"))
query_by(lm \%in\% class || "data.frame" \%in\% class)

# search for a specific name
query_by(is_named("input", "x", "model"))
query_by(name == "input" || name == "x" || name == "model")
}

}
