% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{reduce_steps}
\alias{reduce_steps}
\alias{remove_step}
\alias{verify_step}
\title{Operations on steps.}
\usage{
reduce_steps(s, dots, store)

remove_step(s, id)

verify_step(step, dots, parent_env, store)
}
\arguments{
\item{s}{Graph of steps, see [experiment::graph_to_steps].}

\item{dots}{Keep nodes that meet these conditions, see [lazyeval::dots].}

\item{store}{Object store to read tags from [storage::os_read_tags].}

\item{id}{Identifier of object to be removed.}

\item{step}{Step to be verified.}

\item{parent_env}{Evaluate `dots` in environment descending from this one.}
}
\value{
`reduce_steps` returns a reduced steps graph derived from `s`.

`remove_step` returns a reduced steps graph derived from `s`.

`verify_step` returns `TRUE` if `step` meets the criteria and
        `FALSE` otherwise.
}
\description{
`reduce_steps` reduces the graph of steps according to
conditions specified in `dots`.

`remove_step` removes a single step from graph of steps
`s` and updates links accordingly.

`verify_step` checks whether `step` meets the condition
defined in `dots`. It evaluates `dots` in an environment that has
`parent_env` as its parent. Object tags are read from `store`.
}
