% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.r
\name{read_commit}
\alias{read_commit}
\title{Read a commit from storage.}
\usage{
read_commit(st, id)
}
\arguments{
\item{st}{Storage to read from.}

\item{id}{Commit identifier.}
}
\description{
Commit read from storage has
}

