% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{recorded}
\alias{recorded}
\alias{is_recorded}
\alias{print.recorded}
\alias{summary.recorded}
\alias{plot.recorded}
\alias{as_tibble.recorded}
\alias{.DollarNames.recorded}
\alias{$.recorded}
\title{Sequence of R commands.}
\usage{
is_recorded(x)

\method{print}{recorded}(x, ...)

\method{summary}{recorded}(x, ...)

\method{plot}{recorded}(x, ...)

as_tibble.recorded(x, ...)

\method{.DollarNames}{recorded}(x, pattern = "")

\method{$}{recorded}(x, i)
}
\description{
A \code{recorded} object represents a part of the
history of R session as recorded by the \link{tracker}. It can be printed
to the console, plotted via a \code{Shiny} widget but, most importantly,
it can be \emph{replayed} in a variety of ways, including arbitrary
substitutions of objects created or stored in any of the past states
of R session.
}
\examples{
\dontrun{
x <- 1
y <- x + 1
v <- y + 100
z <- (x + 2) * z

s <- tracker_sequence()
print(s)
summary(s)
plot(s)

# replay the whole sequence replacing x in the initial commit (a) with 100
t <- s(a:x = 100)
print(t)

# show the object x in commit a
t$a$x

# export all data as a tibble
as_tibble(t)
}
}
