% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.r
\name{replace_literals}
\alias{replace_literals}
\title{Replace literals in an expression.}
\usage{
replace_literals(expression, literals)
}
\arguments{
\item{expression}{A \code{language} object.}

\item{literals}{A named \code{list} of atomic values.}
}
\value{
\code{expression} with respective symbols replaced with their values.
}
\description{
Replace all occurences of a given symbol with its value. A values needs
to be a single-element atomic vector.
}

