% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{restore}
\alias{restore}
\title{Restore a snapshot from history.}
\usage{
restore(id)
}
\arguments{
\item{id}{\code{commit} or object identifier, a SHA1 string (\strong{long id})
or its first 8 characters (\strong{short id}).}
}
\description{
Restores a historical \link{commit} from commit or object \code{id}:
\itemize{
\item commit \code{id} brings back that specific commit
\item object \code{id} brings back the earlies commit where that
object can be found, which should be the commit where that
object has been created
}
}
