% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.r
\name{restore_commit}
\alias{restore_all_commits}
\alias{restore_commit}
\title{Read a commit from storage.}
\usage{
restore_commit(st, id)

restore_all_commits(st)
}
\arguments{
\item{st}{Storage to read from.}

\item{id}{Commit identifier.}
}
\value{
A \code{commit} object.

A \code{commit_set} object.
}
\description{
Commit read from storage has

Restore all commits.
}

