% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{results}
\alias{results}
\title{A set (list) of results.}
\usage{
results(res)
}
\arguments{
\item{res}{A \code{list} of objects to wrap in a \code{results} class.}
}
\description{
A set (list) of results.
}
