% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.r
\name{search_for_assignment}
\alias{search_for_assignment}
\title{Recursively search for assignment to a given variable.}
\usage{
search_for_assignment(name, expression)
}
\arguments{
\item{name}{Variable name.}

\item{expression}{Expression to search in.}
}
\value{
An \code{expression} that produces the variable value or
        \code{NULL} if could not find it.
}
\description{
Recursively search for assignment to a given variable.
}

