% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{data}
\name{state}
\alias{state}
\title{State object.}
\format{An object of class \code{environment} of length 5.}
\usage{
state
}
\description{
\describe{
  \item{last}{Stores handles to all objects in the global namespace
    at the time of the last commit.}
  \item{tracking}{whether we are in the tracking state}
  \item{old_prompt}{prompt as set when loading the package}
  \item{stash}{local. file-system-based, object cache}
}
}
\keyword{datasets}

