% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{is_steps}
\alias{is_steps}
\alias{plot.steps}
\alias{print.steps}
\title{Interactive history.}
\usage{
is_steps(x)

\method{plot}{steps}(x, ...)

\method{print}{steps}(x, ...)
}
\arguments{
\item{x}{The \code{steps} history object to be printed or viewed in \code{plot}; the object to be tested in \code{is_steps}.}

\item{...}{Extra arguments for printing/plotting.}
}
\description{
\code{is_steps} verifies if the given object is a valid
\code{steps} structure.

\code{plot.steps} open an interactive history viewer.

\code{plot.steps} open an interactive history viewer.
}
