% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.r
\name{storage}
\alias{storage}
\title{Initialize/create a new/open existing storage.}
\usage{
storage(path, .create = FALSE)
}
\arguments{
\item{path}{Path to the folder.}

\item{.create}{Create the storage if it is not under \code{path}.}
}
\description{
Initialize/create a new/open existing storage.
}

