% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.r
\name{store_commit}
\alias{store_commit}
\title{Convert environment into a commit.}
\usage{
store_commit(env, parent_id, history, storage)
}
\arguments{
\item{env}{Environment to be stored.}

\item{parent_id}{ID of the parent commit.}

\item{history}{History line that triggered this store operation.}

\item{storage}{Storage object.}
}
\value{
ID of the new commit in \code{storage}.
}
\description{
A commit in the filesystem is just another regular object with two
mandatory tags: \code{class == 'commit'} and \code{parent} that
points to the parent commit. Thus, reading a commit from storage
does not differ from reading any other object.
}

