% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{store_environment}
\alias{store_environment}
\alias{store_plot}
\alias{extract_parents}
\alias{names_to_ids}
\alias{strip_object}
\title{Removes references to environments.}
\usage{
store_environment(store, env, expr)

store_plot(store, plot, env, expr, ids)

extract_parents(env, expr)

names_to_ids(what, mapping)

strip_object(obj)
}
\arguments{
\item{obj}{Object to be processed.}
}
\value{
\code{obj} with environment references replaced by \code{emptyenv()}
}
\description{
\code{names_to_ids} is a simple helper function that maps
names to identifiers and makes sure all requested names are mapped.
Important details is to keep the vector of parents named: only this
way an object/plot can be re-computed; without names there is no way
to match object ids into the associated expression.

Some objects (e.g. formula, lm) store references to environments
in which they were created. This function replaces each such reference
with a reference to \code{emptyenv()}.
}
\details{
As much as possible, this function tries not to make any copies of
the original data. This is because the address of the object might
be used to determine whether object's identifier needs to be computed
which might be a costly operation.
}
