% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.r
\name{store_object}
\alias{store_object}
\title{Add an object to a storage.}
\usage{
store_object(st, id, obj, tags = list(), .overwrite = FALSE)
}
\arguments{
\item{st}{Storage object.}

\item{id}{Add object under this identifier.}

\item{obj}{An object to be added.}

\item{tags}{Tags that describe \code{obj}.}

\item{.overwrite}{Overwrite if already exists.}
}
\value{
Invisibly a hash of \code{obj}.
}
\description{
Add an object to a storage.
}

