% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{svg_equal}
\alias{svg_equal}
\title{Compare two SVG images.}
\usage{
svg_equal(a, b)
}
\arguments{
\item{a}{First SVG image.}

\item{b}{Second SVG image.}
}
\value{
`TRUE` if SVGs are the same plot-wise.
}
\description{
SVG images are processed in this package as base64-encoded, XML text
data. When produced, certain differences are introduced in XML
attributes that have no effect on the final plots. That is why,
however, SVG plots need to be compared graphically and not textually.
This function produces a thumbnail of each SVG image and then
compares the raster graphic.
}
