% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{tracked_sequence}
\alias{tracked_sequence}
\alias{is_tracked_sequence}
\alias{print.tracked_sequence}
\alias{summary.tracked_sequence}
\alias{plot.tracked_sequence}
\alias{as_tibble.tracked_sequence}
\alias{.DollarNames.tracked_sequence}
\alias{$.tracked_sequence}
\title{Sequence of R commands.}
\usage{
is_tracked_sequence(x)

\method{print}{tracked_sequence}(x, ...)

\method{summary}{tracked_sequence}(x, ...)

\method{plot}{tracked_sequence}(x, ...)

as_tibble.tracked_sequence(x, ...)

\method{.DollarNames}{tracked_sequence}(x, pattern = "")

\method{$}{tracked_sequence}(x, i)
}
\description{
A \code{tracked_sequence} object represents a part of the
history of R session as recorded by the \link{tracker}. It can be printed
to the console, plotted via a \code{Shiny} widget but, most importantly,
it can be \emph{replayed} in a variety of ways, including arbitrary
substitutions of objects created or stored in any of the past states
of R session.
}
\examples{
\dontrun{
x <- 1
y <- x + 1
v <- y + 100
z <- (x + 2) * z

s <- tracker_sequence()
print(s)
summary(s)
plot(s)

# replay the whole sequence replacing x in the initial commit (a) with 100
t <- s(a:x = 100)
print(t)

# show the object x in commit a
t$a$x

# export all data as a tibble
as_tibble(t)
}
}
