% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\docType{data}
\name{tracker}
\alias{tracker}
\alias{print.tracker}
\title{R session tracker.}
\format{An object of class \code{tracker} of length 2.}
\usage{
tracker

\\method{print}{tracker}(x, ...)
}
\description{
The \code{tracker} object controls whether and how the history
of R session is written to the object repository. The \code{tracker} object
can be used in two ways:
\itemize{
\item printing the object shows the status of the tracker
\item the dollar sign \code{$} gives access to a number of methods and objects
that control specific aspects of \code{tracker}'s behavior
}

Methods and objects exposed via \code{tracker}:
\itemize{
\item \code{tracker$on()} enables history tracking in the current R session
\item \code{tracker$off()} disables history tracking (reverts the \code{on()} method)
\item \code{tracker$attach(path)} attaches
}
}
\keyword{datasets}
