% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{tracker_replay}
\alias{tracker_replay}
\title{Repeat a sequence of commands.}
\usage{
tracker_replay(...)
}
\description{
Repeat a sequence of commands.
}
\examples{
\dontrun{
# initial sequence
x <- 1
y <- x + 2
z <- y ** 2
w <- sqrt(y)

# alterations: explicit value
tracker_replay(x = 2)
# alterations: replace the same object
x <- 2
tracker_replay(x)
# alterations: name substitution
v <- 3
tracker_replay(x = v)
# alterations: only some objects; w is not replayed
tracker_replay(output(z), replace(x = 4))

# show all branches created in those replays
tracker$branch
}

}
