% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.r
\name{update_prompt}
\alias{update_prompt}
\title{Updates the prompt according to \code{state$tracking}.}
\usage{
update_prompt(on_off)
}
\description{
Updates the prompt according to \code{state$tracking}.
}

