% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{commit}
\alias{commit}
\alias{is_commit}
\alias{print.commit}
\title{Creates a new commit object.}
\usage{
commit(artifacts, expression, parent, id = NA_character_)

is_commit(x)

\method{print}{commit}(x, simple = FALSE, header = TRUE, ..., store)
}
\arguments{
\item{artifacts}{A named \code{list} of object ids.}

\item{expression}{Expression associated with this commit.}

\item{parent}{Identifier of the parent commit.}

\item{id}{Identifier of this commit.}

\item{x}{Object to be tested or printed.}

\item{simple}{Show simplified printout.}

\item{header}{If extended output (\code{simple} is \code{FALSE}), print a header line.}

\item{...}{Additional parameters to control printout.}

\item{store}{Optionally, the \link[storage:object_store]{storage::object_store} that holds the
commit; defaults in the internal store of the R session.}
}
\description{
Creates a new commit object.
}
