% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{internal_object_store}
\alias{internal_object_store}
\alias{discover_object_store}
\alias{prepare_object_store}
\alias{create_stash}
\alias{reattach_to_store}
\title{Manage object stores.}
\usage{
discover_object_store(path = getwd())

prepare_object_store(path, .silent = !interactive())

create_stash(path = file.path(tempdir(), "experiment-stash"))

reattach_to_store(state, store, env, .global = "abort",
  .silent = !interactive())
}
\arguments{
\item{path}{Path to be examined.}

\item{.silent}{Reduce the verbosity to warnings and errors.}

\item{state}{The \link{internal_state} object or a testing mock object.}

\item{store}{A \link[storage:object_store]{storage::object_store} object.}

\item{env}{The \link{globalenv} or testing mock environment.}

\item{.global}{Action to take when reading the \link{commit} into \code{env}.}
}
\value{
\code{discover_object_store} returns a \code{character} vector whose
elemnets are paths to existing object stores.

\code{prepare_object_store} returns an \link[storage:object_store]{storage::object_store} object.
}
\description{
Manage object stores.

\code{discover_object_store} starts with with given \code{path}
and searches for a configuration file (\emph{not implemented yet}) or an
existing object store that \code{experiment} can use for the current R
session. If a store can be found, an \link[storage:object_store]{storage::object_store} object
is returned; otherwise, a \code{NULL} value is returned.

\code{prepare_object_store} opens an existing object store
(via \code{discover_object_store}) or creates a new, temporary one.

\code{create_stash} creates an empty object store under the
given \code{path}.

\code{reattach_to_store} makes \code{store} the current object
store and determines which \code{commit} present in that store should
become the current \emph{parent commit} (in \emph{git} known as \code{HEAD}). It
is used only in \link{tracker_on} but it is separate from it for
testing purposes.
}
