% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\docType{data}
\name{internal_state}
\alias{internal_state}
\title{Global state of the tracker.}
\format{An object of class \code{environment} of length 4.}
\usage{
internal_state
}
\description{
\describe{
\item{tracking}{whether we are in the tracking state}
\item{old_prompt}{prompt as set when loading the package}
\item{stash}{local, file-system-based, object cache}
}
}
\keyword{datasets}
