% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{reduce_steps}
\alias{reduce_steps}
\alias{remove_step}
\alias{verify_step}
\title{Operations on steps.}
\usage{
reduce_steps(s, dots, store)

remove_step(s, id)

verify_step(step, dots, parent_env, store)
}
\arguments{
\item{s}{Graph of steps, see \link[experiment:graph_to_steps]{experiment::graph_to_steps}.}

\item{dots}{Keep nodes that meet these conditions, see \link[lazyeval:dots]{lazyeval::dots}.}

\item{store}{Object store to read tags from \link[storage:os_read_tags]{storage::os_read_tags}.}

\item{id}{Identifier of object to be removed.}

\item{step}{Step to be verified.}

\item{parent_env}{Evaluate \code{dots} in environment descending from this one.}
}
\value{
\code{reduce_steps} returns a reduced steps graph derived from \code{s}.

\code{remove_step} returns a reduced steps graph derived from \code{s}.

\code{verify_step} returns \code{TRUE} if \code{step} meets the criteria and
\code{FALSE} otherwise.
}
\description{
\code{reduce_steps} reduces the graph of steps according to
conditions specified in \code{dots}.

\code{remove_step} removes a single step from graph of steps
\code{s} and updates links accordingly.

\code{verify_step} checks whether \code{step} meets the condition
defined in \code{dots}. It evaluates \code{dots} in an environment that has
\code{parent_env} as its parent. Object tags are read from \code{store}.
}
