% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{result}
\alias{result}
\title{A single result.}
\usage{
result(object, id)
}
\arguments{
\item{object}{Wrap object into a \code{result} class.}

\item{id}{Object identifier in \link[storage:object_store]{storage::object_store}.}
}
\description{
A single result.
}
