% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{simulate_london_meters}
\alias{simulate_london_meters}
\title{Simulations and examples.}
\usage{
simulate_london_meters(overwrite = FALSE)
}
\arguments{
\item{overwrite}{If current stash contains objects, setting \code{overwrite}
to \code{TRUE} will remove them prior to running simulation.}
}
\description{
\code{simulate_london_meters} loads and examines a subset
of \strong{London meters} data; see \link{energy} and the introductory vignette.
}
\details{
These functions populate sessions' history cache with a complete
history of data exploration.
}
