% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.r
\name{task_callback}
\alias{task_callback}
\title{A callback run after each top-level expression is evaluated.
 
From \link{\code{addTaskCallback}}: if the data argument was
specified in the call to addTaskCallback, that value is given as the
fifth argument.}
\usage{
task_callback(expr, result, successful, printed)
}
\arguments{
\item{expr}{Expression for the top-level task.}

\item{result}{Result of the top-level task.}

\item{successful}{A logical value indicating whether it was
successfully completed or not (always \code{TRUE} at present).}

\item{printed}{A logical value indicating whether the result was
printed or not.}
}
\value{
A logical value indicating whether to keep this function in
        the list of active callback.
}
\description{
A callback run after each top-level expression is evaluated.
 
From \link{\code{addTaskCallback}}: if the data argument was
specified in the call to addTaskCallback, that value is given as the
fifth argument.
}

