% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{tracker_replay}
\alias{tracker_replay}
\alias{tracker_session}
\alias{tracker_session_}
\alias{tracker_sequence}
\alias{tracker_sequence_}
\title{Work with commands recorded in the current R session.}
\usage{
tracker_replay(..., store, last_id)

tracker_session()

tracker_session_(store)

tracker_sequence()

tracker_sequence_(store, last_id)
}
\description{
\code{tracker_replay} repeats a sequence of commands.

\code{tracker_session} works with the current R session and
shows its summary; It returns a \code{"tracked"} object which can be printed
(console) and plotted (Shiny widget).

\code{tracker_session_} is a non-interactive alternative
which can work with historical R sessions.

\code{tracker_sequence} shows a sequence of commands leading
from the most current to the very beginning of the current R session.
It can be printed (console) and plotted (Shiny widget).

\code{tracker_sequence_} is a non-interactive alternative
which can work with historical R sessions.
}
\examples{
\dontrun{
# initial sequence
x <- 1
y <- x + 2
z <- y ** 2
w <- sqrt(y)

# alterations: explicit value
tracker_replay(x = 2)
# alterations: replace the same object
x <- 2
tracker_replay(x)
# alterations: name substitution
v <- 3
tracker_replay(x = v)
# alterations: only some objects; w is not replayed
tracker_replay(output(z), replace(x = 4))

# show all branches created in those replays
tracker$branch
}

}
