% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{tracker_attach}
\alias{tracker_attach}
\alias{tracker_on}
\alias{tracker_off}
\title{Control object tracking in your R session.}
\usage{
tracker_attach(path = file.path(getwd(), "project-store"),
  .global = "abort")

tracker_on()

tracker_off()
}
\arguments{
\item{path}{Where to locate the object store (see \link[storage:object_store]{storage::object_store}).}

\item{.global}{How to handle \link{globalenv} when it is not empty.}
}
\description{
\code{tracker_on}, \code{tracker_off} and \code{tracker_attach} are
also available as methods of the \link{tracker} object. \code{tracker_on} turns
the tracking mode on, which is signaled by a new prompt, \code{[tracked] >}.
\code{tracker_off} reverses that and turns the tracking mode off. They both
need to be preceeded by a call to \code{tracker_attach} which controls with
file-system object store (see: \link[storage:object_store]{storage::object_store}) is used for
persistent storage.
}
\details{
\code{tracker_attach} attaches to an \link[storage:object_store]{storage::object_store}) if
one can be found under \code{path}. Otherwise, if \code{path} points to a
non-existing directory whose parent directory does exist, then the
top-level directory in \code{path} is created with a new object store in it.

When an existing object store is found under \code{path} and it is not
empty, that is, it contains artifacts and \link{commit}s from previous R
sessions, the current R session is set as a continuation of one of those
\code{commit}s. However, if the current \emph{global environment} (see
\link{globalenv}) is not empty, it needs to be replaced or merged with one
of the \code{commit}s found in the object store. To that extent, the
\code{.global} argument is consulted. It can take one of the following
values:
\itemize{
\item \code{"ask"} - the default, shows the user a dialog with the three other
choices
\item \code{"abort"} - aborts the tracking if \code{globalenv} is not empty
\item \code{"replace"} - replace the contents of \code{globalenv} with the chosen
commit
\item \code{"merge"} - merge the contents of \code{globalenv} with the chosen commit;
this creates a new commit in the process, which is immediately written
back to the object store
}

When tracking is enabled a task callback installed via
\link{addTaskCallback}. It is used to examine the contents of the
\emph{global environment}  each time an R command is successfully executed.
}
\examples{
\dontrun{
# if no object store exists, a new one is created under the
# default "project-store" directory located in the current
# working directory
tracker_attach()

# as above, but the new directory is "my-store"
tracker_attach("my-store")
}

}
