% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{tracking_on}
\alias{tracking_on}
\alias{tracking_off}
\title{Turn tracking on or off}
\usage{
tracking_on()

tracking_off()
}
\description{
\code{tracking_on} turns the tracking mode on. This is
signaled by a new prompt, \code{[tracked] >}. When tracking is enabled,
a special callback installed via \link{addTaskCallback}, is used to examine
the contents of the \emph{global environment} each time an R command is
successfully executed.

\code{tracking_off} reverses the effect of \code{tracking_on}. It
removes the callback and brings back the original value of that R
session's prompt.
}
